% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\docType{data}
\name{constant}
\alias{constant}
\title{Default Input of Relative Risk}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
constant
}
\description{
This dataset gives default input of Relative Risk, however
could be updated latter by the user
\describe{
\item{cvd}{Relative Risk for cardiovascular disease}
\item{diabetes}{Relative Risk for diabetes}
\item{cancer}{Relative Risk for cancer}
\item{copd_asthma}{Relative Risk for chronic obstructive pulmonary disease or asthma}
\item{osteoarthritis}{Relative Risk for osteoarthritis}
\item{hypertension}{Relative Risk for hypertension}
\item{kidney}{Relative Risk for kidney diseases}
\item{g_p_l}{Relative Risk for gallbladder, liver, and pancreatic diseases}
\item{stroke}{Relative Risk for strokes}
}
}
\keyword{datasets}
