% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{data_load}
\alias{data_load}
\title{data_load function}
\usage{
data_load()
}
\value{
a list(dataframe) of pop (population), gdp (global gdp),
mi (median income), bmi (body mass index), disab (disability rate),
employ (employment rate), med_cost (medical conditions cost),
med_prev (medical conditions prevalence cost),
natl_med_prev  (national medical conditions prevalence),
rel_risk (relative risk), benefits, and insurance
}
\description{
The function "data_load" would load all critical values in a returned list format
}
\examples{
raw_data <- data_load()
population <- raw_data$pop
}
