% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_produce.R
\name{data_produce}
\alias{data_produce}
\title{data_produce function}
\usage{
data_produce(rr = c())
}
\arguments{
\item{rr}{the relative risks of diseases -- Cardiovascular disease, diabetes, cancer, Chronic obstructive pulmonary disease or asthma, osteoarthritis, hypertension, kidney diseases, (Gallbladder, Liver, Pancreatic) diseases, and strokes.}
}
\value{
a list (dataframe) of constant, obesity_cost_disease, obesity_cost_full, and obesity_cost_national_summary
}
\description{
The data_produce function would generate the essential four tables that
concerns obesity including
1. Relative Risks (constant): Relative risks for a given disease group with a risk factor of obesity.
2. Disease Cost (obesity_cost_disease): Supplementary output with all variables related to individual disease groups in a given state and year.
3. Full Cost (obesity_cost_full): Complete output with all variables used to make cost calculations, as well as cost calculations in a given state and year.
4. National Summary (obesity_cost_national_summary): National summary cost calculations in a given year.
}
\examples{
new_data <- data_produce(rr = c(1,2,3,4,5,6,7,8,9.1))
cnst <- new_data$constant
}
