% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity_cost_disease.R
\docType{data}
\name{obesity_cost_disease}
\alias{obesity_cost_disease}
\title{Relevant Data for Obesity, Cost, and Diseases}
\format{
An object of class \code{data.frame} with 10350 rows and 8 columns.
}
\usage{
obesity_cost_disease
}
\description{
This dataset gives supplementary output with all variables
related to individual disease groups in a given state and year.
\describe{
\item{State}{state of interest}
\item{Year}{year of interest}
\item{pi_it}{obesity prevalence in state i and year t}
\item{cause}{disease group}
\item{rr_j}{relative risk of disease group j on obesity}
\item{psi_jt}{national cost of disease group j in year t}
\item{rho_jit}{population-attributable risk percent of disease group j in state i and year t}
\item{DC_jit}{direct cost for disease group j in state i and year t}
}
}
\keyword{datasets}
