% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity_cost_national_summary.R
\docType{data}
\name{obesity_cost_national_summary}
\alias{obesity_cost_national_summary}
\title{National summary cost calculations in a given year}
\format{
An object of class \code{data.frame} with 23 rows and 9 columns.
}
\usage{
obesity_cost_national_summary
}
\description{
National summary cost calculations in a given year
\describe{
\item{Year}{year of interest}
\item{DC_t}{direct cost in year t}
\item{M_t}{excess mortality cost in year t}
\item{A_t}{absenteeism cost in year t}
\item{D_t}{disability cost in year t}
\item{IC_t}{indirect cost in year t}
\item{TC_t}{total cost in year t}
\item{p_t}{total population in year t}
\item{pi_t}{average obesity rate in year t}
}
}
\keyword{datasets}
