% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_risk_fun.R
\name{rel_risk_fun}
\alias{rel_risk_fun}
\title{rel_risk_fun function}
\usage{
rel_risk_fun(rr)
}
\arguments{
\item{rr}{the relative risks of diseases -- Cardiovascular disease, diabetes, cancer, Chronic obstructive pulmonary disease or asthma, osteoarthritis, hypertension, kidney diseases, (Gallbladder, Liver, Pancreatic) diseases, and strokes.}
}
\value{
a list (dataframe) of relative risks
}
\description{
The "rel_risk_fun" could update the relative risks (or the constants) when crucial data needs updating
}
\examples{
key <- rel_risk_fun(rr = c(1,2,3,4,5,6,7,8,1.2))
key$rr
}
